+++ appinfo/info.xml
@@ -37,7 +37,7 @@ See [README](https://github.com/rello/audioplayer/blob/master/README.md) file fo
 	<screenshot>https://raw.githubusercontent.com/rello/audioplayer/master/screenshots/audioplayer_screen.png</screenshot>
 	<dependencies>
 		<php min-version="7.0" max-version="7.4"/>
-		<nextcloud min-version="19" max-version="21"/>
+		<nextcloud min-version="20" max-version="21"/>
 	</dependencies>
 	<settings>
 		<admin-section>OCA\audioplayer\Settings\AdminSection</admin-section>
